create custom persisted data hook
