+++ src/components/App.js
@@ -1,9 +1,10 @@
-import React from 'react';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import React from "react";
+import { BrowserRouter as Router, Route } from "react-router-dom";
 
-import GlobalStyles from './GlobalStyles';
-import Home from './Home';
-import Game from './Game';
+import GlobalStyles from "./GlobalStyles";
+import { items } from "../data";
+import Home from "./Home";
+import Game from "./Game";
 
 function App(props) {
   return (
